os/pm: Add PM Thread Metrics
This commit extends the functionality of pm_metrics. Now PM Metrics show the thread specific pm metrics. It shows which process causes minimum actual sleep time (in ms) and how frequently the thread wakes up the board.

+++ os/pm/pm_wakehandler.c
@@ -56,7 +56,7 @@ void pm_wakehandler(clock_t missing_tick, pm_wakeup_reason_code_t wakeup_src)
 {
 	irqstate_t flags = enter_critical_section();
 #ifdef CONFIG_PM_METRICS
-	pm_metrics_update_wakehandler(missing_tick, wakeup_src);
+	pm_metrics_update_wakehandler((uint32_t)missing_tick, wakeup_src);
 #endif
 	pmllvdbg("wakeup source code = %d\n", wakeup_src);
 	pmllvdbg("missing_tick: %llu\n", missing_tick);
